+++ module/helpers/active-effect-config.sheet.js
@@ -76,9 +76,19 @@ export class ArM5eActiveEffectConfig extends ActiveEffectConfig {
           context.selectedSubtypes[idx]
         ].default == "boolean"
       ) {
-        tmpSubTypes[context.selectedSubtypes[idx]].bool = true;
+        tmpSubTypes[context.selectedSubtypes[idx]].isBool = true;
       } else {
-        tmpSubTypes[context.selectedSubtypes[idx]].bool = false;
+        tmpSubTypes[context.selectedSubtypes[idx]].isBool = false;
+      }
+
+      if (
+        typeof ACTIVE_EFFECTS_TYPES[context.selectedTypes[idx]].subtypes[
+          context.selectedSubtypes[idx]
+        ].default == "string"
+      ) {
+        tmpSubTypes[context.selectedSubtypes[idx]].isString = true;
+      } else {
+        tmpSubTypes[context.selectedSubtypes[idx]].isString = false;
       }
 
       let withChoice =